resin-image-flasher-unwrap: Do not check image format if not present
The standard flow for this tool is to unwrap a resin-image from a
resin-image-flasher.

On commit cea5fc2a5623 the option of converting it to a different format to
use in emulators was introduced, but the image format was made compulsory
with the script exiting if not provided.

This commit only checks the validity of image format when actually provided.

Change-type: patch
Signed-off-by: Alex Gonzalez <alexg@balena.io>